core, eth, les, miner: move common function to beacon package
